+++ packages/expo-updates/CHANGELOG.md
@@ -45,6 +45,7 @@
 ### 💡 Others
 
 - [iOS] Moved expo-dev-client + expo-updates interop setup out from `application:willFinishLaunchingWithOptions:`. ([#27477](https://github.com/expo/expo/pull/27477) by [@kudo](https://github.com/kudo))
+- [iOS] Added bridgeless support on ExpoReactDelegate. ([#27601](https://github.com/expo/expo/pull/27601) by [@kudo](https://github.com/kudo))
 
 ## 0.24.11 - 2024-02-16